misc: extract some LS constant to a specific file
