+++ src/core/constants/localStorageKeys.ts
@@ -1,0 +1,2 @@
+export const LAST_PRIVATE_VISITED_ROUTE_WHILE_NOT_CONNECTED_LS_KEY = 'lastPrivateVisitedRoute'
+export const ORGANIZATION_LS_KEY_ID = 'currentOrganization'
